squash-merge: N+1 문제 해결

+++ src/main/java/dev/noah/word/response/SearchPostResponse.java
@@ -2,7 +2,13 @@ package dev.noah.word.response;
 
 import java.time.LocalDateTime;
 
-public record SearchPostResponse(long id, String authorName, String authorImageUrl, LocalDateTime createdDate,
-                                 String title, String content, int views, int comments) {
-
+public record SearchPostResponse(
+        long id,
+        String authorName,
+        String authorImageUrl,
+        LocalDateTime createdDate,
+        String title,
+        String content,
+        int views,
+        long comments) {
 }
